+++ packages/wallets/src/wallets/wallet-factory.ts
@@ -1,5 +1,12 @@
 import { WebAuthnP256 } from "ox";
-import type { AdminSignerConfig, ApiClient, CreateWalletParams, GetWalletSuccessResponse } from "../api";
+
+import type {
+    AdminSignerConfig,
+    ApiClient,
+    CreateWalletParams,
+    GetWalletSuccessResponse,
+    RegisterSignerPasskeyParams,
+} from "../api";
 import { WalletCreationError, WalletNotAvailableError } from "../utils/errors";
 import type { Chain } from "../chains/chains";
 import type { InternalSignerConfig, SignerConfigForChain } from "../signers/types";
@@ -165,7 +172,9 @@ export class WalletFactory {
         }
     }
 
-    private async createPasskeyAdminSigner<C extends Chain>(signer: SignerConfigForChain<C>) {
+    private async createPasskeyAdminSigner<C extends Chain>(
+        signer: SignerConfigForChain<C>
+    ): Promise<RegisterSignerPasskeyParams> {
         if (signer.type !== "passkey") {
             throw new Error("Signer is not a passkey");
         }