+++ apps/auth/node-and-express-server/CHANGELOG.md
@@ -1,5 +1,23 @@
 # crossmint-auth-node
 
+## 1.1.43
+
+### Patch Changes
+
+- @crossmint/server-sdk@1.2.32
+
+## 1.1.42
+
+### Patch Changes
+
+- @crossmint/server-sdk@1.2.31
+
+## 1.1.41
+
+### Patch Changes
+
+- @crossmint/server-sdk@1.2.30
+
 ## 1.1.40
 
 ### Patch Changes