+++ apps/payments/create-react-app/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @crossmint/client-sdk-react-ui-starter
 
+## 1.3.117
+
+### Patch Changes
+
+- Updated dependencies [c7a5d49]
+  - @crossmint/client-sdk-react-ui@2.2.13
+
+## 1.3.116
+
+### Patch Changes
+
+- @crossmint/client-sdk-react-ui@2.2.12
+
+## 1.3.115
+
+### Patch Changes
+
+- Updated dependencies [174d730]
+  - @crossmint/client-sdk-react-ui@2.2.11
+
+## 1.3.114
+
+### Patch Changes
+
+- @crossmint/client-sdk-react-ui@2.2.10
+
 ## 1.3.113
 
 ### Patch Changes