+++ apps/wallets/quickstart-devkit/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @crossmint/wallets-quickstart-devkit
 
+## 0.1.34
+
+### Patch Changes
+
+- Updated dependencies [c7a5d49]
+  - @crossmint/client-sdk-react-ui@2.2.13
+
+## 0.1.33
+
+### Patch Changes
+
+- @crossmint/client-sdk-react-ui@2.2.12
+
+## 0.1.32
+
+### Patch Changes
+
+- Updated dependencies [174d730]
+  - @crossmint/client-sdk-react-ui@2.2.11
+
+## 0.1.31
+
+### Patch Changes
+
+- @crossmint/client-sdk-react-ui@2.2.10
+
 ## 0.1.30
 
 ### Patch Changes