+++ packages/client/verifiable-credentials/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @crossmint/client-sdk-verifiable-credentials
 
+## 3.4.49
+
+### Patch Changes
+
+- @crossmint/client-sdk-base@1.5.11
+
+## 3.4.48
+
+### Patch Changes
+
+- Updated dependencies [fa8da3f]
+  - @crossmint/common-sdk-base@0.8.4
+  - @crossmint/client-sdk-base@1.5.10
+
+## 3.4.47
+
+### Patch Changes
+
+- Updated dependencies [174d730]
+  - @crossmint/client-sdk-base@1.5.9
+  - @crossmint/common-sdk-base@0.8.3
+
 ## 3.4.46
 
 ### Patch Changes