+++ packages/common/base/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @crossmint/common-sdk-base
 
+## 0.8.4
+
+### Patch Changes
+
+- fa8da3f: fix: world-chain string
+
+## 0.8.3
+
+### Patch Changes
+
+- 174d730: checkout: order metadata
+
 ## 0.8.2
 
 ### Patch Changes